Fix Vim statusline file-content encoding
Explicitly compare &fileencoding to an empty string to force conversion
to a boolean. Vim script does not coerce a string to a boolean which
means that &encoding was always shown as the file-content encoding in
the status line.
Use a hardcoded "utf-8" string instead of the value of the &encoding
option. The &encoding option has nothing to do with file-content
encoding.