Use virtcol('.') in Vim visual "Q" and "."
Use virtcol('.') instead of getcurpos()[4] in Vim visual "Q" and "."
mappings as virtcol('.') returns the virtual column of the cursor
whereas getcurpos()[4] returns curswant. Note that curswant can be very
large after moving to the the end of the line with "$" which is simply
unnecessary when passed to the GetExprNormal() function. The actual
virtual column of the cursor is desired instead.