Improve style of Vim Quickfix command
Improve the style of fzf entries for the Vim Quickfix command by using
whitespace instead of a semicolon to delimit the error number.
Use a regex as the fzf delimiter to match the error number and semicolon
delimiters.
Update how the selected fzf entry is converted to a usable error number.